damage Fixed

+++ Assets/Scripts/Players/PlayerHealthBar.cs
@@ -7,13 +7,13 @@ public class PlayerHealthBar : MonoBehaviour
 {
     public Slider _slider;
 
-    public void SetMaxHealth(float _health)
+    public void SetMaxHealth(int _health)
     {
         _slider.maxValue = _health;
         _slider.value = _health;
     }
 
-    public void ChangeHealthBar(float _health)
+    public void ChangeHealthBar(int _health)
     {
         _slider.value = _health;
     }
